requirements.txt
requests==2.31.0
python-decouple==3.8
ipython==8.13.2
django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
whitenoise==6.4.0
gunicorn==20.1.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
dj-database-url==2.0.0
celery==5.2.7
redis==4.5.5
django-celery-beat==2.5.0
python-telegram-bot==13.15  # last sync version